- major refactor into monorepo - refactored bake-library into external module @azbake/core.    - This sets up the refactor work needed to support external ingredients
