+++ .gitignore
@@ -3,5 +3,5 @@
 .history
 
 ### NPM
-node_modules/*
-dist/*
+node_modules
+dist
