+++ core/.npmignore
