+++ core/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@azbake/core",
+  "description": "core bake library to integrate 3rd party ingredients into the system",
+  "main": "dist/index.js",
+  "author": "HCHB",
+  "license": "MIT",
+  "version": "0.1.5",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/HomecareHomebase/azure-bake.git"
+  },
+  "bugs": {
+    "url": "https://github.com/HomecareHomebase/azure-bake/issues"
+  },
+  "homepage": "https://github.com/HomecareHomebase/azure-bake",
+  "scripts": {
+    "upload": "tsc -p ./src && npm --no-git-tag-version version patch && npm publish --access public",
+    "compile": "tsc -p ./src",
+    "watch": "tsc -w -p ./src",
+    "debug": "node --inspect-brk=0.0.0.0:9229 dist/index.js",
+    "start": "node dist/index.js",
+    "test": "mocha --timeout 30000 -r ts-node/register test/**/*.ts"
+  },
+  "peerDependencies": {
+    "colors": ">= 1.3.2",
+    "@azure/ms-rest-nodeauth": ">= 0.8.3"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@types/node": "^10.12.18"
+  }
+}
