+++ core/src/deployment-context.ts
@@ -1,9 +1,9 @@
-import { BakePackage, IBakeRegion, IBakeConfig, IBakeEnvironment } from "./bake-loader";
+import { IBakePackage, IBakeRegion, IBakeConfig, IBakeEnvironment } from "./bake-interfaces";
 import { Logger } from "./logger";
 import { TokenCredentialsBase } from "@azure/ms-rest-nodeauth";
 
 export class DeploymentContext {
-    constructor(auth: TokenCredentialsBase, bake : BakePackage, region: IBakeRegion, logger: Logger) {
+    constructor(auth: TokenCredentialsBase, bake : IBakePackage, region: IBakeRegion, logger: Logger) {
 
         this._package = bake
         this._region = region
@@ -11,7 +11,7 @@ export class DeploymentContext {
         this._auth = auth
     }
 
-    _package: BakePackage
+    _package: IBakePackage
     _region: IBakeRegion
     _logger: Logger
     _auth: TokenCredentialsBase
