+++ core/src/index.ts
@@ -1,0 +1,10 @@
+import {BakeEval} from './eval'
+import {BakeVariable} from './bake-variable'
+import * as functions from './functions'
+import {Logger} from './logger'
+import {DeploymentContext} from './deployment-context'
+import {IBakeAuthentication, IBakeConfig, IBakeEnvironment, IBakePackage, IBakeRegion, IIngredient, IIngredientProperties} from './bake-interfaces'
+
+export {BakeEval, BakeVariable, functions, DeploymentContext,Logger,
+    IBakeAuthentication, IBakeConfig, IBakeEnvironment, IBakePackage, IBakeRegion, IIngredient, IIngredientProperties
+}
