+++ core/src/tsconfig.json
