--- src/bake-library/index.ts
@@ -1,5 +1,0 @@
-import {BakeEval} from './eval'
-import {BakeVariable} from './bake-variable'
-import * as functions from './functions'
-
-export {BakeEval, BakeVariable, functions}
