+++ system/src/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "es6",
+    "module": "commonjs",
+    "declaration": true,
+    "outDir": "../dist",
+    "strict": true,
+    "stripInternal": true,
+    "sourceMap": true
+  },
+  "include": ["**/*.ts"],
+}
